Use Popper instead of Modal
